@@ -13,17 +13,18 @@
     "@react-navigation/stack": "^6.4.1",
     "babel-plugin-module-resolver": "^5.0.2",
     "expo": "^51.0.36",
+    "expo-font": "~12.0.10",
     "expo-status-bar": "~1.12.1",
     "nativewind": "^4.1.10",
     "react": "18.2.0",
     "react-native": "0.74.5",
-    "react-native-reanimated": "^3.15.4"
+    "react-native-reanimated": "~3.10.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.45",
     "tailwindcss": "^3.3.2",
-    "typescript": "^5.1.3"
+    "typescript": "~5.3.3"
   },
   "private": true
 }
